Circuits for hidden services don't need to be exit enabled circuits
Split up tests for tunnel and hidden community
