+++ Tribler/community/tunnel/routing.py
@@ -13,7 +13,7 @@ class Circuit(object):
     """ Circuit data structure storing the id, state and hops """
 
     def __init__(self, circuit_id, goal_hops=0, first_hop=None, proxy=None,
-                 ctype=CIRCUIT_TYPE_DATA, callback=None, required_exit=None, 
+                 ctype=CIRCUIT_TYPE_DATA, callback=None, required_endpoint=None,
                  mid=None, info_hash=None):
         """
         Instantiate a new Circuit data structure
@@ -44,7 +44,7 @@ class Circuit(object):
         self.proxy = proxy
         self.ctype = ctype
         self.callback = callback
-        self.required_exit = required_exit
+        self.required_endpoint = required_endpoint
         self.mid = mid
         self.hs_session_keys = None
         self.info_hash = info_hash
