+++ Tribler/community/tunnel/tunnel_community.py
@@ -184,7 +184,6 @@ class TunnelExitSocket(DatagramProtocol):
 class TunnelSettings(object):
 
     def __init__(self, install_dir=None, tribler_session=None):
-        self.circuit_length = 3
         self.crypto = TunnelCrypto()
         self.socks_listen_ports = range(1080, 1085)
 
@@ -255,7 +254,6 @@ class TunnelCommunity(Community):
         self.circuits_needed = {}
         self.exit_candidates = {}
         self.notifier = None
-        self.made_anon_session = False
         self.selection_strategy = RoundRobin(self)
         self.stats = defaultdict(int)
         self.creation_time = time.time()
@@ -405,16 +403,12 @@ class TunnelCommunity(Community):
 
     def do_circuits(self):
         for circuit_length, num_circuits in self.circuits_needed.items():
-            num_to_build = num_circuits - sum([1 for c in self.circuits.itervalues() if c.goal_hops == circuit_length])
-            self._logger.debug("want %d circuits of length %d", num_to_build, circuit_length)
-
+            num_to_build = num_circuits - sum([1 for _ in self.data_circuits(circuit_length).itervalues()])
+            self._logger.debug("want %d data circuits of length %d", num_to_build, circuit_length)
             for _ in range(num_to_build):
-                try:
-                    if not self.create_circuit(circuit_length):
-                        break
-                except:
-                    self._logger.exception("Error creating circuit while running do_circuits")
-
+                if not self.create_circuit(circuit_length):
+                    self._logger.debug("circuit creation of %d circuits failed, no need to continue" % num_to_build)
+                    break
         self.do_remove()
 
     def tunnels_ready(self, hops):
@@ -460,63 +454,65 @@ class TunnelCommunity(Community):
                 self.exit_candidates.pop(pubkey)
                 logging.debug("Removed candidate from exit_candidates dictionary")
 
-    def create_circuit(self, goal_hops, ctype=CIRCUIT_TYPE_DATA, callback=None, max_retries=0, required_exit=None, info_hash=None):
-        assert required_exit is None or isinstance(required_exit, tuple), type(required_exit)
-        assert required_exit is None or len(required_exit) == 3, required_exit
+    def create_circuit(self, goal_hops, ctype=CIRCUIT_TYPE_DATA, callback=None, max_retries=0, required_endpoint=None, info_hash=None):
+        assert required_endpoint is None or isinstance(required_endpoint, tuple), type(required_endpoint)
+        assert required_endpoint is None or len(required_endpoint) == 3, required_endpoint
         retry_lambda = first_hop = None
 
         if max_retries > 0:
-            retry_lambda = lambda h = goal_hops, t = ctype, c = callback, r = max_retries - 1, e = required_exit: \
+            retry_lambda = lambda h = goal_hops, t = ctype, c = callback, r = max_retries - 1, e = required_endpoint: \
                 self.create_circuit(h, t, c, r, e)
 
-        if not required_exit:
-            self._logger.debug("Look for exit node to set as required_exit for this circuit")
-            # Each circuit's exit node should be a verified connectable exit node peer chosen by the circuit initiator
+        if not required_endpoint:
             for c in self.dispersy_yield_verified_candidates():
                 pubkey = c.get_member().public_key
-                exit_candidate = self.exit_candidates[pubkey]
-                if exit_candidate.become_exit:
-                    self._logger.debug("Valid exit candidate found for this circuit")
-                    required_exit = (c.sock_addr[0], c.sock_addr[1], pubkey)
-
-                    # Stop looking for a better alternative if the exit-node is not used for exiting in another circuit
+                if ctype == CIRCUIT_TYPE_DATA:
+                    self._logger.debug("Look for an exit node to set as required_endpoint for this circuit")
+                    exit_candidate = self.exit_candidates[pubkey]
+                    if exit_candidate.become_exit:
+                        self._logger.debug("Valid exit candidate found for this circuit")
+                        required_endpoint = (c.sock_addr[0], c.sock_addr[1], pubkey)
+                        break
+                else:
+                    self._logger.debug("Try to find a connectable node to set as required_endpoint for this circuit")
+                    required_endpoint = (c.sock_addr[0], c.sock_addr[1], pubkey)
                     if self.candidate_is_connectable(c):
-                        self._logger.debug("Exit node is connectable and not used in other circuits, that's prefered")
+                        self._logger.debug("Valid required_endpoint found for this circuit, stop looking further")
                         break
 
-        # If the number of hops is 1, it should immediately be the required_exit hop.
-        if goal_hops == 1 and required_exit:
+        # If the number of hops is 1, it should immediately be the required_endpoint hop.
+        if goal_hops == 1 and required_endpoint:
             self._logger.debug("Associate firsthop with a candidate and member object")
-            first_hop = Candidate((required_exit[0], required_exit[1]), False)
-            first_hop.associate(self.get_member(public_key=required_exit[2]))
+            first_hop = Candidate((required_endpoint[0], required_endpoint[1]), False)
+            first_hop.associate(self.get_member(public_key=required_endpoint[2]))
         else:
-            # Otherwise, look for a first hop that is not used before.
+            self._logger.debug("Look for a first hop that is not used before.")
             hops = set([c.first_hop for c in self.circuits.values()])
             for c in self.dispersy_yield_verified_candidates():
                 if (c.sock_addr not in hops) and self.crypto.is_key_compatible(c.get_member()._ec) and \
-                   (not required_exit or c.sock_addr != tuple(required_exit[:2])):
+                   (not required_endpoint or c.sock_addr != tuple(required_endpoint[:2])):
                     first_hop = c
                     break
 
-        if not required_exit:
+        if not required_endpoint:
             if retry_lambda:
-                self._logger.info("could not create circuit, no available exit-nodes found, will retry in 5 seconds.")
+                self._logger.debug("could not create circuit, no available exit-nodes found, will retry in 5 seconds.")
                 self.register_task(retry_lambda, reactor.callLater(5, retry_lambda))
             else:
-                self._logger.info("could not create circuit, no available exit-nodes.")
+                self._logger.debug("could not create circuit, no available exit-nodes.")
             return False
 
         if not first_hop:
             if retry_lambda:
-                self._logger.info("could not create circuit, no available relay for first hop, will retry in 5 seconds.")
+                self._logger.debug("could not create circuit, no available relay for first hop, will retry in 5 seconds.")
                 self.register_task(retry_lambda, reactor.callLater(5, retry_lambda))
             else:
-                self._logger.info("could not create circuit, no available relay for first hop.")
+                self._logger.debug("could not create circuit, no available relay for first hop.")
             return False
 
         circuit_id = self._generate_circuit_id(first_hop.sock_addr)
         circuit = Circuit(circuit_id, goal_hops, first_hop.sock_addr, self, ctype, callback,
-                          required_exit, first_hop.get_member().mid.encode('hex'), info_hash)
+                          required_endpoint, first_hop.get_member().mid.encode('hex'), info_hash)
 
         self.request_cache.add(CircuitRequestCache(self, circuit, retry_lambda))
 
@@ -524,7 +520,7 @@ class TunnelCommunity(Community):
         circuit.unverified_hop.address = first_hop.sock_addr
         circuit.unverified_hop.dh_secret, circuit.unverified_hop.dh_first_part = self.crypto.generate_diffie_secret()
 
-        self._logger.info("creating circuit %d of %d hops. First hop: %s:%d", circuit_id,
+        self._logger.debug("creating circuit %d of %d hops. First hop: %s:%d", circuit_id,
                           circuit.goal_hops, first_hop.sock_addr[0], first_hop.sock_addr[1])
 
         self.circuits[circuit_id] = circuit
@@ -644,6 +640,10 @@ class TunnelCommunity(Community):
             sock_addr = self.exit_sockets[circuit_id].sock_addr
             self.send_destroy(Candidate(sock_addr, False), circuit_id, reason)
             self._logger.debug("destroy_exit_socket %s %s", circuit_id, sock_addr)
+
+    def data_circuits(self, hops=None):
+        return {cid: c for cid, c in self.circuits.items()
+                if c.ctype == CIRCUIT_TYPE_DATA and (hops is None or hops == len(c.hops))}
 
     def active_data_circuits(self, hops=None):
         return {cid: c for cid, c in self.circuits.items()
@@ -790,13 +790,13 @@ class TunnelCommunity(Community):
         if circuit.state == CIRCUIT_STATE_EXTENDING:
             ignore_candidates = [self.crypto.key_to_bin(hop.public_key) for hop in circuit.hops] + \
                 [self.my_member.public_key]
-            if circuit.required_exit:
-                ignore_candidates.append(circuit.required_exit[2])
+            if circuit.required_endpoint:
+                ignore_candidates.append(circuit.required_endpoint[2])
 
             become_exit = circuit.goal_hops - 1 == len(circuit.hops)
-            if become_exit and circuit.required_exit:
+            if become_exit and circuit.required_endpoint:
                 # Set the required exit according to the circuit setting (e.g. for linking e2e circuits)
-                host, port, pub_key = circuit.required_exit
+                host, port, pub_key = circuit.required_endpoint
                 extend_hop_public_bin = pub_key
                 extend_hop_addr = (host, port)
 
@@ -1106,20 +1106,20 @@ class TunnelCommunity(Community):
         for message in messages:
             circuit_id = message.payload.circuit_id
             cand_sock_addr = message.candidate.sock_addr
-            self._logger.error("Got destroy from %s for circuit %s", message.candidate, circuit_id)
+            self._logger.debug("Got destroy from %s for circuit %s", message.candidate, circuit_id)
 
             if circuit_id in self.relay_from_to:
                 self.remove_relay(circuit_id, "Got destroy", True, (circuit_id, cand_sock_addr))
 
             elif circuit_id in self.exit_sockets:
-                self._logger.error("Got an exit socket %s %s", circuit_id, cand_sock_addr)
+                self._logger.debug("Got an exit socket %s %s", circuit_id, cand_sock_addr)
                 if cand_sock_addr != self.exit_sockets[circuit_id].sock_addr:
                     self._logger.error("%s not allowed send destroy", cand_sock_addr)
                     continue
                 self.remove_exit_socket(circuit_id, "Got destroy")
 
             elif circuit_id in self.circuits:
-                self._logger.error("Got a circuit %s %s", circuit_id, cand_sock_addr)
+                self._logger.debug("Got a circuit %s %s", circuit_id, cand_sock_addr)
                 if cand_sock_addr != self.circuits[circuit_id].first_hop:
                     self._logger.error("%s not allowed send destroy", cand_sock_addr)
                     continue
@@ -1169,8 +1169,8 @@ class TunnelCommunity(Community):
         self.send_data([Candidate(sock_addr, False)], u'data', packet)
 
     def exit_data(self, circuit_id, sock_addr, destination, data):
-        if not self.become_exitnode():
-            self._logger.error("Dropping data packets, refusing to be an exit node")
+        if not self.become_exitnode() and not TunnelConversion.could_be_dispersy(data):
+            self._logger.error("Dropping data packets, refusing to be an exit node for data")
         elif circuit_id in self.exit_sockets:
             if not self.exit_sockets[circuit_id].enabled:
                 # We got the correct circuit_id, but from a wrong IP.
